fix(renderer): Update braille renderer for qrcode-rust2 repo
This adjusts the braille renderer to fit with conventions and constraints of the
qrcode-rust2 repository.

Copyright texts are added, unsafe code is removed, and required traits are
derived. I added Shun Sakai in copyright for the benchmark because the benchmark
is almost a copy-paste of the unicode benchmark. This is so that the performance
can be compared.

To increase readability a little bit I also moved any byte related code to a new
`BrailleByte` struct.